refactor(shared): lift commitSave + editor hooks out of /src/data
commitSave (pure save→onSaved chain) moves data/vehicles/api → src/utils so
both vehicle features can share it without a cross-feature import. Extract the
byte-identical EditingContext dirty-lift and ?selected= close-slider blocks from
VehicleDetails into src/hooks/useLiftEditorDirty + useCloseSliderParam.
